requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
cryptography==41.0.3
Django==4.2.5
django-environ==0.11.2
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.4
msgpack==1.0.5
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
redis==5.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
